Support asynchronous methods for @Timed and @Exception (#26)
`@Timed` and `@Exception` when used to intercept methods that return an
implementation of `CompletionStage<?>` will wait for the future to
resolve before stopping the timer and recording any failures,
respectively. This enables the metrics to more accurately capture the
operation performed by the method even if the method itself returns
early and defers the actual work asynchronously.

Additionally, imports and copyright notices have been updated across the
project.